+++ chat/server/index.js
@@ -19,7 +19,7 @@ const server = createServer(app)
 
 const io = new Server(server, {
   cors: {
-    origin: 'http://localhost:3000',
+    origin: '*',
     methods: ['GET', 'POST'],
     credentials: true
   }
