--- src/chat/client/package.json
@@ -1,47 +1,0 @@
-{
-  "name": "client",
-  "private": true,
-  "type": "module",
-  "version": "0.0.0",
-  "scripts": {
-    "dev": "vite",
-    "build": "run-p type-check \"build-only {@}\" --",
-    "preview": "vite preview",
-    "build-only": "vite build",
-    "type-check": "vue-tsc --build --force",
-    "lint": "eslint . --fix"
-  },
-  "dependencies": {
-    "@mdi/font": "7.4.47",
-    "core-js": "^3.37.1",
-    "roboto-fontface": "*",
-    "socket.io-client": "^4.8.1",
-    "vue": "^3.4.31",
-    "vue-emoji-picker": "^1.0.3",
-    "vuetify": "^3.6.14"
-  },
-  "devDependencies": {
-    "@eslint/js": "^9.14.0",
-    "@tsconfig/node22": "^22.0.0",
-    "@types/node": "^22.9.0",
-    "@vitejs/plugin-vue": "^5.1.4",
-    "@vue/eslint-config-typescript": "^14.1.3",
-    "@vue/tsconfig": "^0.5.1",
-    "eslint": "^9.14.0",
-    "eslint-plugin-vue": "^9.30.0",
-    "npm-run-all2": "^7.0.1",
-    "pinia": "^2.1.7",
-    "sass": "1.77.8",
-    "sass-embedded": "^1.77.8",
-    "typescript": "~5.6.3",
-    "unplugin-auto-import": "^0.17.8",
-    "unplugin-fonts": "^1.1.1",
-    "unplugin-vue-components": "^0.27.2",
-    "unplugin-vue-router": "^0.10.0",
-    "vite": "^5.4.10",
-    "vite-plugin-vue-layouts": "^0.11.0",
-    "vite-plugin-vuetify": "^2.0.3",
-    "vue-router": "^4.4.0",
-    "vue-tsc": "^2.1.10"
-  }
-}
