--- src/chat/server/package.json
@@ -1,28 +1,0 @@
-{
-  "name": "server",
-  "version": "1.0.0",
-  "description": "",
-  "main": "index.js",
-  "type": "module",
-  "scripts": {
-    "start": "node server",
-    "build": "tsc",
-    "dev": "nodemon server"
-  },
-  "keywords": [],
-  "author": "",
-  "license": "ISC",
-  "dependencies": {
-    "cors": "^2.8.5",
-    "express": "^5.1.0",
-    "moment": "^2.30.1",
-    "mysql2": "^3.14.1",
-    "socket.io": "^4.8.1",
-    "sqlite": "^5.1.1",
-    "sqlite3": "^5.1.7"
-  },
-  "devDependencies": {
-    "nodemon": "^3.1.9",
-    "unplugin-auto-import": "^19.1.2"
-  }
-}
